gSOAP 2.8.81 stable
git-svn-id: https://svn.code.sf.net/p/gsoap2/code@163 06ac4634-b970-40ce-84dc-3611c6d995e3

+++ gsoap/samples/xml-rpc-json/json.cpp
@@ -406,7 +406,7 @@ int json_recv(struct soap *soap, struct value *v)
     }
     case '['/*']'*/:
     {
-      size_t i;
+      int i;
       if (++soap->level > soap->maxlevel)
         return soap->error = SOAP_LEVEL;
 #ifdef __cplusplus
@@ -423,7 +423,7 @@ int json_recv(struct soap *soap, struct value *v)
       if (c == /*'['*/']')
         return SOAP_OK;
       soap_unget(soap, c);
-      for (i = 0; i < soap->maxoccurs; i++)
+      for (i = 0; i < (int)soap->maxoccurs; i++)
       {
 #ifdef __cplusplus
         if (json_recv(soap, v->operator[](i)))
